refactor: pkg to internal
